docs(icons): add missing icons documentation
