+++ playwright/e2e/element-examples/static.spec.ts
@@ -54,9 +54,9 @@ test('si-electron-titlebar/si-electron-titlebar', ({ si }) => si.static());
 test('si-electron-titlebar/si-fixed-height-layout-side-panel', ({ si }) => si.static());
 test('si-empty-state/si-empty-state', ({ si }) => si.static());
 test('si-footer/si-footer', ({ si }) => si.static());
-test('si-icon/si-icon-composite', ({ si }) => si.static());
 test('si-icon/si-icon', ({ si }) => si.static({ skipAriaSnapshot: true }));
-test('si-icon/si-icon-next', ({ si }) => si.static({ skipAriaSnapshot: true }));
+test('si-icon/si-icon-composite', ({ si }) => si.static());
+test('si-icon/si-icon-legacy', ({ si }) => si.static({ skipAriaSnapshot: true }));
 test('si-icon-status/si-icon-status', ({ si }) => si.static());
 test('si-info-page/si-info-page', ({ si }) => si.static());
 test('si-info-page/si-info-page-stacked-icon', ({ si }) => si.static());
