+++ playwright/snapshots/static.spec.ts-snapshots/si-icon--si-icon-legacy-element-examples-chromium-dark-linux.png
@@ -1,0 +1,3 @@
+version https://git-lfs.github.com/spec/v1
+oid sha256:374590e56b9cfe2d9772df6c2c1e99142c3101ccd0703a8483d4c68c21ebed0a
+size 8034
