+++ playwright/snapshots/static.spec.ts-snapshots/si-icon--si-icon-legacy-element-examples-chromium-light-linux.png
@@ -1,0 +1,3 @@
+version https://git-lfs.github.com/spec/v1
+oid sha256:52e152719cc008db7f72676726e9d03dabb56634da4c155173f1dfb2c9856283
+size 8188
