--- playwright/snapshots/static.spec.ts-snapshots/si-icon--si-icon-next-element-examples-chromium-dark-linux.png
@@ -1,3 +1,0 @@
-version https://git-lfs.github.com/spec/v1
-oid sha256:d0253a3cfd3404029ecea6039e79a2b52b5fcf8dd5f2a606fb8864b8bc49f193
-size 8690
