--- playwright/snapshots/static.spec.ts-snapshots/si-icon--si-icon-next-element-examples-chromium-light-linux.png
@@ -1,3 +1,0 @@
-version https://git-lfs.github.com/spec/v1
-oid sha256:6c5269b84fa2106d730f5a23f91050463f4881bfaed649fa876ada77946a0f66
-size 8840
