+++ src/app/examples/si-icon/si-icon-legacy.html
@@ -1,0 +1,4 @@
+<si-icon icon="element-user" class="icon" />
+<si-icon icon="element-plant" class="si-display-xl" />
+<si-icon icon="element-security" class="text-secondary si-display-xl" />
+<si-icon icon="element-alchemy-flask" class="text-primary si-display-xl" />
