--- src/app/examples/si-icon/si-icon-next.html
@@ -1,19 +1,0 @@
-<!-- Preferred approach: using the object to add the icon. This is typesafe and ensure that the icon was imported. -->
-<si-icon [icon]="icons.elementUser" />
-<si-icon class="icon" [icon]="icons.elementUser" />
-<si-icon class="si-display-xl" [icon]="icons.elementUser" />
-<!-- This will load the SVG as well, as we imported element-plus, but less save. Only use the string directly for dynamic values from the backend. -->
-<si-icon icon="element-2d-editor" />
-<si-icon class="icon" icon="element-2d-editor" />
-<si-icon class="si-display-xl" icon="element-2d-editor" />
-
-<!-- Stacked in template -->
-<span class="icon icon-stack">
-  <si-icon class="status-danger" [icon]="icons.elementAlarmFilled" />
-  <si-icon class="text-secondary" [icon]="icons.elementAlarmTick" />
-</span>
-
-<!-- Icons are hidden in the a11y tree. Add an aria-label if no other descriptive text is provided.-->
-<button class="btn btn-primary btn-circle btn-xs" type="button" aria-label="Cancel">
-  <si-icon [icon]="icons.elementCancel" />
-</button>
