+++ src/app/examples/si-icon/si-icon.ts
@@ -3,12 +3,30 @@
  * SPDX-License-Identifier: MIT
  */
 import { Component } from '@angular/core';
-import { SiIconComponent } from '@siemens/element-ng/icon';
+import {
+  addIcons,
+  element2dEditor,
+  elementAlarmFilled,
+  elementAlarmTick,
+  elementCancel,
+  elementUser,
+  SiIconModule,
+  SiIconComponent
+} from '@siemens/element-ng/icon';
 
 @Component({
   selector: 'app-sample',
-  imports: [SiIconComponent],
+  imports: [SiIconModule, SiIconComponent],
   templateUrl: './si-icon.html',
   host: { class: 'p-5' }
 })
-export class SampleComponent {}
+export class SampleComponent {
+  // addIcons returns a map of all names added to the library for typesafe use in the template.
+  icons = addIcons({
+    elementUser,
+    elementCancel,
+    element2dEditor,
+    elementAlarmFilled,
+    elementAlarmTick
+  });
+}
